Move manifest and change detection helpers out of cli code
