+++ src/ossature/cli/commands/audit.py
@@ -32,7 +32,12 @@ from ossature.audit.interfaces import (
     infer_interface_from_smd,
     propagate_to_smd_dependents,
 )
-from ossature.audit.manifest import create_manifest, read_manifest, write_manifest
+from ossature.audit.manifest import (
+    check_and_update_manifest,
+    create_manifest,
+    get_changed_spec_ids,
+    write_manifest,
+)
 from ossature.audit.planner import (
     collect_orphaned_output_files,
     generate_plan,
@@ -54,7 +59,6 @@ from ossature.models.audit import (
     SpecAuditReport,
 )
 from ossature.models.smd import SMDSpec
-from ossature.models.vmd import VMDSpec
 from ossature.parsers.amd import AMDParseError, parse_amd_file
 from ossature.parsers.smd import SMDParseError, parse_smd_file
 from ossature.parsers.vmd import VMDParseError
@@ -174,86 +178,6 @@ def _build_amd_file_map(
     for amd_file, amd in zip(amd_files, parsed_amds, strict=True):
         result.setdefault(amd.spec_id, []).append(str(amd_file.relative_to(spec_dir)))
     return result
-
-
-def check_and_update_manifest(
-    console: Console,
-    config: OssatureConfig,
-    smd_files: list[Path],
-    amd_files: list[Path],
-    vmd_files: list[Path] | None = None,
-) -> tuple[list[str] | None, Manifest]:
-    """Returns (changed source keys or None if unchanged, current manifest).
-
-    Brief input hashes from a prior manifest are carried forward so brief
-    regeneration can compare against them after fix cycles.
-    """
-    config.metadata_path.mkdir(parents=True, exist_ok=True)
-    manifest_path = config.metadata_path / "manifest.toml"
-
-    old: Manifest | None = None
-    if manifest_path.exists():
-        console.log("Reading existing manifest")
-        old = read_manifest(manifest_path)
-        if not old:
-            console.log("Malformed manifest. Disregarding.")
-
-    new_manifest = create_manifest(
-        config=config,
-        smd_files=smd_files,
-        amd_files=amd_files,
-        vmd_files=vmd_files,
-        brief_inputs=old.brief_inputs if old else None,
-        project_brief_input=old.project_brief_input if old else "",
-    )
-
-    if old is not None:
-        mismatched = new_manifest.diff(other=old)
-        if mismatched:
-            console.log("[red]Manifest changed")
-            for source in mismatched:
-                console.log(f"  {source} has changed")
-            write_manifest(new_manifest, filename=manifest_path)
-            console.log("Manifest updated")
-            return mismatched, new_manifest
-        else:
-            console.log("[green]Manifest unchanged")
-            return None, new_manifest
-
-    write_manifest(new_manifest, filename=manifest_path)
-    console.log("[green]Manifest written")
-    return list(new_manifest.sources.keys()), new_manifest
-
-
-def get_changed_spec_ids(
-    changed_files: list[str],
-    smd_files: list[Path],
-    amd_files: list[Path],
-    parsed_smds: list[SMDSpec],
-    parsed_amds: list[AMDSpec],
-    config: OssatureConfig,
-    vmd_files: list[Path] | None = None,
-    parsed_vmds: list[VMDSpec] | None = None,
-) -> set[str]:
-    """Maps changed manifest source keys to spec IDs."""
-    if "ossature.toml" in changed_files:
-        return {smd.spec_id for smd in parsed_smds}
-
-    file_to_spec: dict[str, str] = {}
-
-    for smd_file, smd in zip(smd_files, parsed_smds, strict=True):
-        key = str(smd_file).replace(str(config.root), ".")
-        file_to_spec[key] = smd.spec_id
-
-    for amd_file, amd in zip(amd_files, parsed_amds, strict=True):
-        key = str(amd_file).replace(str(config.root), ".")
-        file_to_spec[key] = amd.spec_id
-
-    for vmd_file, vmd in zip(vmd_files or [], parsed_vmds or [], strict=True):
-        key = str(vmd_file).replace(str(config.root), ".")
-        file_to_spec[key] = vmd.spec_id
-
-    return {file_to_spec[f] for f in changed_files if f in file_to_spec}
 
 
 def generate_and_write_briefs(
